pyproject.toml
[tool.poetry]
name = "RitsukiBot"
version = "0.1.0"
description = "RitsukiBot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot-adapter-onebot = "^2.1"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot-plugin-imageutils = "^0.1.8.1"
nb-cli = "^0.6.7"
Pillow = "^9.2.0"
numpy = "1.23.1"
ujson = "^5.4.0"
peewee = "^3.15.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
